\docType{methods}
\name{as.qw-methods}
\alias{as.qw-methods}
\alias{as.qw,numeric,character,character,numeric,character,character,character,character,character,missing-method}
\title{Methods for Function \code{as.qw}}
\description{
  Valid conversion methods for function \code{as.qw}.
}
\section{Methods}{
  \describe{

  \item{list("signature(values = \"numeric\", remark.codes
  = \"character\", value.codes = \"character\",
  reporting.level = \"numeric\", reporting.method =
  \"character\", reporting.units = \"character\",
  analyte.method = \"character\", analyte.name =
  \"character\", unique.code = \"character\", value2 =
  \"missing\")")}{ All of the data and meta data are
  required for objects of class "qw."}}
}
\keyword{manip}
\keyword{methods}

