% Generated by roxygen2 (4.0.2): do not edit by hand
\name{censCor}
\alias{censCor}
\title{Correlation}
\usage{
censCor(x, y, Full = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{a left-censored data object.}

\item{y}{a left-censored data object.}

\item{Full}{how to compute  the mean and standard deviation of \code{x} and \code{y}.
See \bold{Details}.}

\item{na.rm}{logical, remove missing values before computing the correlation?}
}
\value{
A vector with these names:
\item{cor}{ the correlation between \code{x} and \code{y}.}
\item{mnx}{ the mean of \code{x}.}
\item{sdx}{ the standard deviation of \code{x}.}
\item{mny}{ the mean of \code{y}.}
\item{sdy}{ the standard deviation of \code{y}.}
\item{cx}{ the proportion of censored values of \code{x}.}
\item{cy}{ the proportion of censored values of \code{y}.}
\item{cxy}{ the proportion of censored values common to \code{x} and \code{y}.}
\item{n}{ the number of observations.}
\item{ll0}{ the log likelihood for cor=0}
\item{llcor}{ the log likelihood for cor=cor}
}
\description{
Compute the correlation between left-censored data using MLE
}
\details{
\code{Full} may be either logical or a numeric vector. If \code{Full} is \code{TRUE},
then estimate the means and standard deviations for \code{x} and \code{y}.
If \code{Full} is \code{FALSE}, use the intial maximum likelihood estimate for those
statistics. Otehrwise \code{Full} can be a named vector containing \code{mnx}, the mean
for \code{x}; \code{sdx}, the standard deviation for \code{x}; \code{mny}, the mean
for \code{y}; \code{sdy}, the standard deviation for \code{y}. \code{Full} can be set
to \code{FALSE} if the optimization fails at large censoring levels or to improve
processing speed for large sample sizes.
}
\examples{
# Simple no censoring
set.seed(450)
tmp.X <- rnorm(25)
tmp.Y <- tmp.X/2 + rnorm(25)
cor(tmp.X, tmp.Y)
censCor(tmp.X, tmp.Y)
# Some censoring
censCor(as.lcens(tmp.X, -1), as.lcens(tmp.Y, -1))
}
\references{
Lyles and others, Biometrics 2001
}
\keyword{censored}
\keyword{summary}

