% Generated by roxygen2 (4.0.2): do not edit by hand
\name{censMove.1}
\alias{censMove.1}
\title{Censored Maintenance of Variance Extension, Type 1}
\usage{
censMove.1(formula, data, subset, na.action, distribution = "normal")
}
\arguments{
\item{formula}{a formula with a single response variable on the left and asingle
explanatory variable on the right. See \bold{Details}.}

\item{data}{the data to search for the variables in \code{formula}.}

\item{subset}{an expression to select a subset of the data.}

\item{na.action}{what to do with missing values.}

\item{distribution}{the distribution of the data, either "normal," "lognormal,"
or "commonlog." See \bold{Details}.}
}
\value{
An object of class "move.1" as decribed in the documentation for \code{move.1}.
Except, there are no \code{fitted.values} or \code{residuals} components, the values of
\code{x} and \code{y} include estimated values (See \bold{Note}), and components \code{cx}
and \code{cy}, which indicate left-censoring are included.
}
\description{
Calculates the Maintenance Of Variance Extension, Type 1 (MOVE.1) for record
extension by fitting a Line of Organic Correlation (LOC) regression model to
left-censored data.
}
\details{
The left- or right-hand side of the formula may be any numeric variable, just as with
\code{move.1} or a variable of class "lcens"\cr
If distribution is "normal," then the data in x and y are assumed to have a
bivariate normal distribution. Otherwise, they are assumed to have a bivariate
log-normal distribution and a logarithmic transform is applied to both x and y
before analysis. The natural logarithm is used if distribution is "lognormal" and
the commmon logarithm is used if distribution is "commonlog."
}
\note{
The components \code{x} and \code{y} are the actual values if the data are
uncensored. If \code{y} is censored and \code{x} is not censored, then \code{y} is the
expected value of \code{y}, given its censored value and the value of \code{x} and likewise
for \code{x} censored and \code{y} not censored. If both are cesnored, then the values
are computed from their expected values given the censored value without respect to
any condition of the other.
}
\examples{
\dontrun{
library(USGSwsData)
data(IonBalance)
# Build model for non missing Alkalinity
IB.move <- move.1(Anion_sum ~ Cation_sum, data=IonBalance,
 subset=abs(Pct_Diff) < 10)
print(IB.move)
# Compare to censored at
}
}
\references{
will need some.
}
\seealso{
\code{\link{move.1}}
}
\keyword{censored}
\keyword{regression}

