% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censReg.pred.R
\name{censReg_AMLE.pred}
\alias{censReg_AMLE.pred}
\title{Support Function for Censored Regression}
\usage{
censReg_AMLE.pred(fit, XPRED)
}
\arguments{
\item{fit}{the output from \code{censReg_AMLE.fit}.}

\item{XPRED}{a matrix of the prediction variables.}
}
\value{
A list containing the output from the FORTRAN code.
}
\description{
Support function for censored regression to predict values from a 
censored regression model.
}
\seealso{
\code{\link{censReg}},
}

