% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotPlot-censored.R
\docType{methods}
\name{dotPlot-censored}
\alias{dotPlot,mcens-method}
\alias{dotPlot,qw-method}
\alias{dotPlot-censored}
\title{Censored Data Methods for Function \code{dotPlot}}
\usage{
\S4method{dotPlot}{mcens}(x, y, Plot = list(name = "", what = "points", type =
  "solid", width = "standard", symbol = "circle", filled = TRUE, size = 0.09,
  color = "black"), yaxis.orient = "table", yaxis.order = "none",
  yaxis.grid = TRUE, xaxis.log = FALSE, xaxis.range = c(NA, NA),
  ylabels = "full", xlabels = 7, xtitle = deparse(substitute(x)),
  ytitle = "", caption = "", margin = c(NA, NA, NA, NA),
  jitter.y = TRUE, Censored = list(name = "", what = "points", symbol =
  "circle", filled = FALSE, size = 0.09, color = "black", bar = TRUE), ...)

\S4method{dotPlot}{qw}(x, y, Plot = list(name = "", what = "points", type =
  "solid", width = "standard", symbol = "circle", filled = TRUE, size = 0.09,
  color = "black"), yaxis.orient = "table", yaxis.order = "none",
  yaxis.grid = TRUE, xaxis.log = FALSE, xaxis.range = c(NA, NA),
  ylabels = "full", xlabels = 7, xtitle = deparse(substitute(x)),
  ytitle = "", caption = "", margin = c(NA, NA, NA, NA),
  jitter.y = TRUE, Censored = list(name = "", what = "points", symbol =
  "circle", filled = FALSE, size = 0.09, color = "black", bar = TRUE), ...)
}
\arguments{
\item{x}{the x-axis data. Missing values are permitted and not plotted.}

\item{y}{the y-axis data, expected to be be either character or factor.
Missing values are permitted and removed before plotting.}

\item{Plot}{control parameters for uncensored values of the plot, see 
\code{link{setMultiPlot}} and \bold{Details} for details.}

\item{yaxis.orient}{orientation of the y-axis values, must be either "table"
or "grid." "Table" is sorted from top to bottom, "grid" is sorted from
bottom to top.}

\item{yaxis.order}{the order of the y-axis values, must be one of "none,"
"ascending," or "descending."}

\item{yaxis.grid}{draw grid lines?}

\item{xaxis.log}{log-transform the x axis?}

\item{xaxis.range}{set the range of the x-axis}

\item{ylabels}{set up y-axis labels.}

\item{xlabels}{set up x-axis labels.}

\item{xtitle}{x-axis title.}

\item{ytitle}{y-axis title.}

\item{caption}{the figure caption.}

\item{margin}{the parameters of the margin.}

\item{jitter.y}{logical, if \code{TRUE}, then adjust \code{y} values to reduce
overlap for each group, or adjust randomly if no groups. If \code{FALSE}, then
no adjustment is made.}

\item{Censored}{control parameters for censored values of the plot.}

\item{...}{arguments for specific methods.}
}
\value{
Information about the graph.
}
\description{
Produces a dot plot: methods for "mcens" and "qw" data.
}
\details{
The \code{what} component of the \code{Plot} argument must be either
"points" or "none."
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{dotPlot}.
}
\seealso{
\code{\link{setPage}}
}
\keyword{hplot}
\keyword{methods}

