% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ecdfPlot.lcens}
\alias{ecdfPlot.lcens}
\alias{ecdfPlot.qw}
\title{Empirical Distribution Function}
\usage{
\method{ecdfPlot}{lcens}(x, group = NULL, Plot = list(name = "Auto", what =
  "stairstep", type = "solid", width = "standard", symbol = "circle", filled =
  TRUE, size = 0.09, color = "Auto"), xaxis.log = TRUE, xaxis.range = c(NA,
  NA), xlabels = 11, ylabels = 5, ytitle = "Cumulative Probability",
  xtitle = deparse(substitute(x)), caption = "", margin = c(NA, NA, NA,
  NA), ...)

\method{ecdfPlot}{qw}(x, group = NULL, Plot = list(name = "Auto", what =
  "stairstep", type = "solid", width = "standard", symbol = "circle", filled =
  TRUE, size = 0.09, color = "Auto"), xaxis.log = TRUE, xaxis.range = c(NA,
  NA), xlabels = 11, ylabels = 5, ytitle = "Cumulative Probability",
  xtitle = deparse(substitute(x)), caption = "", margin = c(NA, NA, NA,
  NA), ...)
}
\arguments{
\item{x}{the data to plot. Missing values are removed before plotting.}

\item{group}{create groups of the left-censored data vector.}

\item{Plot}{control parameters of the uncensored data plot.}

\item{xaxis.log}{log-transform the x axis?}

\item{xaxis.range}{set the range of the x axis.}

\item{xlabels}{set the x-axis labels. See \code{\link{linearPretty}} for
details.}

\item{ylabels}{set the y-axis labels. See \code{\link{linearPretty}} for
details.}

\item{ytitle}{the y-axis title.}

\item{xtitle}{the x-axis title.}

\item{caption}{the figure caption.}

\item{margin}{the parameters of the margin of the plot area.}

\item{\dots}{any additional arguments needed by specific methods.}
}
\value{
Information about the graph.
}
\description{
Produces a graph of the empirical distribution function: methods
for "lcens" and "qw" data.
}
\examples{
\dontrun{
set.seed(1932)
Xu <- rlnorm(32)
setPage("sq") # required page set up
ecdfPlot(as.lcens(Xu, 1.0))
}
}
\seealso{
\code{\link{probPlot.lcens}}
}
\keyword{hplot}

