% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fillIn}
\alias{fillIn}
\alias{fillIn.lcens}
\alias{fillIn.mcens}
\alias{fillIn.numeric}
\title{Estimate Censored Values}
\usage{
fillIn(x, method, alpha)

\method{fillIn}{numeric}(x, method, alpha)

\method{fillIn}{lcens}(x, method = "ROS", alpha = 0.4)

\method{fillIn}{mcens}(x, method = "ROS", alpha = 0.4)
}
\arguments{
\item{x}{an object of class "lcens." Missing values are ignored.}

\item{method}{the method to use for estimating censored values:
"ROS," "log ROS," "MLE," "log MLE," or "triangular" are valid for left-
censored data; "ROS," "log ROS," "MLE," or "log MLE,"are valid for multiply
censored data; ignored for uncensored data.}

\item{alpha}{the offset fraction to be used; typically in [0,0.5].}
}
\value{
A vector of sorted estimates and actual values with an attribute of the
censoring levels.
}
\description{
Estimates values of censored data.
}
\details{
The methods of Regression on Order Statistics (ROS) and MLE is
explained in Helsel (2012). The "log ROS" first log-transform the data and
back-transforms the estimated values. The triangular method distributes the
censored value assuming a triangular distribution between 0 and the single
detection limit. Quinn and Keogh (2003) describe alternatives to simple
substituion of a single value that sample from an alternate distribution.
The triangular distribution is a reasonable distribution when the percentage
of censored data is relatively small, say less than 30 percent.
}
\examples{
set.seed(5420)
XR <- sort(rnorm(10))
XF <- fillIn(as.lcens(XR, -1)) # censors lowest 3 values
#How'd we do?
rbind(XR, XF)
#Note that this is unusual because all 10 random values were less than the mean!
}
\references{
Helsel, D.R. 2012, Blah blah: some place, Wiley, some pages.\cr
Quinn, G.P. and Keogh, M.J., 2003, Experimental design and data analysis for
biologists, Cambridge University Press, Cambridge, UK, 539 p.
}
\seealso{
\code{\link{as.double.qw}}
}
\keyword{manip}

