% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.na.R
\name{is.na.lcens}
\alias{is.na.lcens}
\alias{is.na.mcens}
\alias{is.na.qw}
\title{Missing Values}
\usage{
\method{is.na}{lcens}(x)

\method{is.na}{mcens}(x)

\method{is.na}{qw}(x)
}
\arguments{
\item{x}{the object to be tested.}
}
\value{
A logical vector of the same length as its argument \code{x},
containing TRUE for those elements marked NA and FALSE otherwise.
}
\description{
Indicate which elements of censored or water-quality data are missing.
}
\examples{
is.na(as.lcens(c(1,3, NA), 2))
}
\keyword{manip}

