\docType{class}
\name{lcens-class}
\alias{lcens-class}
\title{Class \code{"lcens"}}
\description{
  Class "lcens" describes left-censored data.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{as.lcens(values, detlim, censor.codes)}.
}
\examples{
showClass("lcens")
}
\references{
  Lorenz, D.L., in preparation
}
\keyword{classes}

