\docType{methods}
\name{Math-methods}
\alias{Math-methods}
\alias{Math,lcens-method}
\title{Mathematical Transform Methods for \code{lcens} objects}
\description{
  Some limited mathematical methods are possible and
  well-defined for censored or water-quality data.
}
\section{Methods}{
  \describe{

  \item{list("signature(x = \"lcens\")")}{ The functions
  \code{log}, \code{log10}, \code{exp}, and \code{sqrt} are
  defined for objects of class "lcens." } }
}
\keyword{methods}

