\docType{class}
\name{mcens-class}
\alias{mcens-class}
\title{Class \code{"mcens"}}
\description{
  Class "mcens" describes multiply-censored data.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{as.mcens(lower.val, upper.val, censor.codes)}.
}
\examples{
showClass("mcens")
}
\references{
  Lorenz, D.L., in preparation
}
\keyword{classes}

