\name{mdlKM}
\alias{mdlKM}
\title{Estimate Statistics}
\usage{
  mdlKM(x, group)
}
\arguments{
  \item{x}{an object of "lcens" to compute}

  \item{group}{the group variable}
}
\value{
  An object of class "survfit."
}
\description{
  Support function for computing statistics for
  left-censored data.
}
\keyword{misc}

