% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mImputeLessThans.R
\name{mImputeLessThans}
\alias{mImputeLessThans}
\alias{mImputeLessThans.data.frame}
\alias{mImputeLessThans.default}
\title{Impute Censored Values}
\usage{
mImputeLessThans(..., m = 10, initial = c("lrEM", "complete.obs",
  "multRepl"))

\method{mImputeLessThans}{default}(..., m = 10, initial = c("lrEM",
  "complete.obs", "multRepl"))

\method{mImputeLessThans}{data.frame}(..., group = NULL, m = 10,
  initial = c("lrEM", "complete.obs", "multRepl"))
}
\arguments{
\item{m}{the number of multiple imputations desired.}

\item{initial}{the method to use for the initial log-ratio covariance matrix, either
"lrEM" that uses expectation maximization to construct the matrix,
"complete.obs" that uses only the rows with no censored data to construct the matrix, or
"multRepl" that uses simple substitution of censored values to compute the matrix.
See \code{\link[zCompositions]{lrDA}} for details.}

\item{group}{character string, the name of the column in the data frame to indicate
a group for imputation. See \bold{Details}}

\item{\dots}{either a data frame that contains columns of class "qw" or any combination of
individual vectors of class "numeric," "lcens," or "qw." Missing values are
removed before processing.}
}
\value{
A list of length \code{m} containing data frames with imputed censored values. If
\code{group} is used, then the returned object is a tagged list with names corresponding
to the values in the grouping column. Each group contains the \code{m} repititions
of imputed values.
}
\description{
Impute left-censored values multiple times using the data augmentation method.
}
\details{
Imputation of left-censored data requires the assumption of multivariate
log-normality for a single population. If the data represent samples from multiple
populations, then they should be identified by the \code{group} argument. The minimum size
for any group is 3.
}
\seealso{
\code{\link[zCompositions]{lrDA}}, \code{\link{imputeLessThans}}
}

