% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nobs.censReg}
\alias{nobs.censReg}
\title{Extract the Number of Observations from a Fit}
\usage{
\method{nobs}{censReg}(object, ...)
}
\arguments{
\item{object}{an object of class "censReg"---output from \code{censReg}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
The number of observations used to construct the model.
}
\description{
Extracts the number of observations from a censored regression object:
method for "censReg" object.
}
\seealso{
\code{\link{censReg}}
}
\keyword{regression}

