% Generated by roxygen2 (4.0.2): do not edit by hand
\name{percentile.lcens}
\alias{percentile.lcens}
\alias{percentile.qw}
\title{Empirical Cumulative Percent}
\usage{
\method{percentile}{lcens}(x, q, test = ">=", na.rm = TRUE,
  percent = TRUE, exact = TRUE, ...)

\method{percentile}{qw}(x, q, test = ">=", na.rm = TRUE, percent = TRUE,
  exact = TRUE, ...)
}
\arguments{
\item{x}{a vector of class "lcens" or "qw" representing the observed data.}

\item{q}{a vector of numeric quantiles for which the cumulative percent or percent
exceedence is desired.}

\item{test}{a character string indicating the test. The default value, '>=,'
is the percent equalling or exceeding the quantile and '<' would return the
cumulative percent below the quantile.}

\item{na.rm}{a logical value indication whether missing values (NAs) should
be removed or not. If na.rm is \code{FALSE} and there are missing values in
\code{x}, then the result will be NA.  The default value is \code{TRUE}.}

\item{percent}{a logical value indicating whether the result should be
expressed as a percent or proportion. The default value, \code{TRUE}, will
express the result as a percent.}

\item{exact}{logical, if \code{TRUE}, then force strict comparisons. Any
other value is ignored in the current version.}

\item{\dots}{not used, required for method function}
}
\value{
A named vector as long as \code{q} corresponding to the requested
value.
}
\description{
Computes the empirical cumulative percent or percent exceedance of observed data
for specific values: methods for "lcens" and "qw" data.
}
\note{
The stats package contains the \code{ecdf} function that performs a
similar function when \code{test} is "<=."
}
\examples{
set.seed(2342)
Xr <- rlnorm(24)
# The percentage of the observarions greater than or equal to 2
percentile(as.lcens(Xr, 2), 1:5)
# Compare to the raw data:
percentile(Xr, 1:5)
}
\seealso{
\code{\link{ecdf}}
}
\keyword{manip}
\keyword{math}
\keyword{univar}

