% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.censReg.R
\name{plot.summary.censReg}
\alias{plot.summary.censReg}
\title{Diagnostic Plot}
\usage{
\method{plot}{summary.censReg}(x, which = "All", set.up = TRUE, span = 1,
  ...)
}
\arguments{
\item{x}{an object of class "censReg"---output from \code{censReg}}

\item{which}{either "All" or any of a sequence from 1 to 7 indicating which plot, 
see \bold{Details}.}

\item{set.up}{set up the graphics page?}

\item{span}{the span to use for the loess smooth. Set to 0 to suppress.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
The object \code{x} is returned invisibly.
}
\description{
Plot summary AMLE/MLE model diagnostics.
}
\details{
Seven  graphs can be produced by this function. If \code{which} is "All," then all plots are produced.
The argument \code{which} can also be the name of an explanatory variable so that a partial residual 
plot is created for a single variable. Or \code{which} can be any of a sequence of numbers from 1 thorugh 7.
Numeric values for \code{which}:
\enumerate{
\item Fitted vs. Observed
\item Fitted vs. Residual
\item S-L plot
\item A correlogram if dates are available in the model or in the data set
\item Q-normal plot
\item Influence plot
\item Partial residual plots for each explanatory variable
}
}
\note{
The q-normal plot (number 5) plots the working residuals and portrays fits 
for relatively small percentages of censoring very well. For larger percentages of
censoring, the q-normal plot in \code{plot.censReg} (number 2) can portray the 
normality of the residuals more accurately.
}
\seealso{
\code{\link{censReg}}, \code{\link{plot.censReg}}
}
\keyword{hplot}
\keyword{regression}
