% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Power Transforms}
\usage{
pow(x, lambda)
}
\arguments{
\item{x}{any object that can be converted to class "mcens." Must be non negative (zero data are
permitted). Missing values are permitted and missing vlues in the returnned vector.}

\item{lambda}{the power to raise \code{x}. Must be strictly positive.}
}
\value{
An object of class "mcens" that is the result of the requested operations. Each of the range of the
lower and upper values are raised to the power of \code{lambda} and the result divided by \code{lambda}
to approximately maintain scale. But the reciprocal value of \code{lambda} cannot be used to restore the
original values.
}
\description{
The \code{pow} function provides a method for raising censored data to a positive power.
}
\examples{
pow(as.mcens(c(0,2), c(1,2)), 2)
pow(as.mcens(c(0,2), c(1,2)), .5)
}

