% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ppw.test}
\alias{ppw.test}
\title{Test for difference in left-censored samples}
\usage{
ppw.test(x, y, alternative = "two.sided", OBrienFleming = TRUE, data.names)
}
\arguments{
\item{x}{the paired samples to \code{y}. Forced to class "lcens." Missing
values are removed before the analysis.}

\item{y}{the paired values to \code{x}. Missing values are removed before
the analysis.}

\item{alternative}{character string describing the alternative hypothesis.
Must be one of "two.sided," "greater," or "less."}

\item{OBrienFleming}{should the O'Brien-Flemming adjustment be made? See
\bold{Details}.}

\item{data.names}{character string to be used to explain the data. Default
names are derived from the data arguments.}
}
\value{
An object of class "htest" that inherits "ppw."
}
\description{
This function tests for differences in paired left-censored samples from 2
groups.
}
\details{
The O'Brien-Fleming adjustment forces the score for all paired censored
values to be equal to each other. It also forces equality between observed
values that are less than the detection limit in the paried sample by setting
the value to less than the detection limit.
}
\note{
The \code{ppw.test}
function uses the \code{survfit} function. Helsel (2012) describes flipping
the left-censored data so that small values become large and left-censored
values become right-censored values and adapt nonparametric techniques from
survival analysis.\cr
A \code{plot} method is supported for the returned object.
}
\section{Null Hypothesis}{
 The null hypothesis is that the
distributions are not different from one another.
}
\examples{
# Compare uncensored results
set.seed(699)
Xu <- sort(rlnorm(22, 0, 1))
Yu <- sort(rlnorm(22, .425, 1))
# Treat as paired samples
ppw.test(Xu, Yu)
wilcox.test(Xu, Yu, paired=TRUE)
}
\references{
Gehan, E.A., 1965, A generalized Wilcoxon test for comparing
arbitraritly singly censored samples: Biometrika, v. 52, p. 203-223.\cr

Harrington, D.P., and Fleming, T.R., 1982, A class of rank test procedures
for censored survival data: Biometrika, v. 69, p. 553-566.\cr

Helsel, D.R. 2012, Statistics for Censored Environmental Data Using Minitab
and R: New York, Wiley, 324 p.\cr

Peto, R., and Peto, J., 1972, Asymptotically efficient rank invariant test
procedures (with discussion): Journal of the Royal Statistical Society,
Series A v. 135, p. 185-206.\cr

Prentice, R.L. 1978, Linear rank tests with right-censored data: Biometrika,
v. 65, p. 167-179.\cr

Prentice, R.L., and Marke, P., 1979, A qualitative discrepancy between
censored data rank tests: Biometrika, v. 35, p. 861-867.\cr
}
\seealso{
\code{\link{survdiff}}, \code{\link{survfit}},
\code{\link{lcens-class}}
}
\keyword{censored}
\keyword{htest}

