\docType{class}
\name{qw-class}
\alias{qw-class}
\title{Class \code{"qw"}}
\description{
  Class "qw" describes water-quality data.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{as.qw(values, remark.codes, value.codes,
  reporting.level, reporting.method, reporting.units,
  analyte.method, analyte.name, unique.code, value2)}.
}
\examples{
showClass("qw")
}
\references{
  Lorenz, D.L., in preparation
}
\keyword{classes}

