% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwRPD.R
\name{qwRPD}
\alias{qwRPD}
\title{Relative Differences}
\usage{
qwRPD(x, y, incomparables = NA, plotit = FALSE)
}
\arguments{
\item{x}{sample data of class "qw."}

\item{y}{replicate data of class "qw."}

\item{incomparables}{logical, \code{NA} means the RPD for any pair that includes
left-censored values is \code{NA} (default), \code{FALSE} means the RPD for any pair
that includes left-censored values is the minimum possible value, and \code{TRUE} means
the RPD for any pair that includes includes left-censored values returns the 
range of possible values. See \code{Notes}.}

\item{plotit}{logical, if \code{TRUE}, then create a Bland-Altman mean-difference plot 
(Bland and Altman, 1986); otherwise no plot is created. If \code{incomparables} is set to
\code{TRUE}, then the values corresponding to the respective maximums are plotted.}
}
\value{
A vector of the relative percent differences for each paired sample and duplicate.
If \code{incomparables} is \code{TRUE}, then the vector is class "mcens," otherwise the vector
id numeric.
}
\description{
Computes the relative percent differences (RDP) between samples and replicates for data of
class "qw."
}
\note{
This function works differently from other functions that process data of class 
"qw"---it ignores the reporting level information and uses the data as they are recorded. 
The treatment of interval-censored data varies depending on \code{incomparables}. If
\code{incomparables} is set to \code{NA} or \code{FALSE}, then interval-censored data are set to
the midrange, otherwise the data are retained as interval-censored data. Right-censored data
always return NA regardless of \code{incomparables}.

Inverval-censored data are not common and require much additional processing to get
accurate ranges for the RPD. Approximate values are returned in this version.
}
\references{
Bland J.M. and Altman D.G., 1986 Statistical methods for assessing agreement between two 
methods of clinical measurement: Lancet, i, p. 307--310.
}
\seealso{
\code{\link[smwrStats]{rpd}}
}

