% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readNWISqw}
\alias{readNWISqw}
\title{Water-Quality Data}
\usage{
readNWISqw(sites, params = "All", begin.date = "", end.date = "")
}
\arguments{
\item{sites}{a vector of the USGS station identifiers.}

\item{params}{A character string contains the name of a group of parameter
codes, or a vector of pamater codes. See \bold{Details}.}

\item{begin.date}{the earliest date for data, must be a character with the
format "YYYY-mm-dd."}

\item{end.date}{the latest date for data, must be a character with the format
"YYYY-mm-dd."}
}
\value{
A data frame of columns describing water-quality samples organized by
parameter code.
}
\description{
Import discrete sample water-quality data from NWIS web.
}
\details{
Valid groups are "All," "information," "physical," "cations," "anions,"
"nutrients," "microbiological," "biological," "metals," "nonmetals,"
"toxicity," "pesticides," "pcbs," "other organics," "radio chemicals,",
"stable isotopes," "sediment," and "population/community."
}
\note{
Probably need a note.
}
\examples{
\dontrun{
readNWISqw("05330000", "00608") # Ammonia samples from the Minnesota River at Jordan.
}
}
\references{
Lorenz, D.L., 2014, USGSqw OFR.\cr See information about discrete
samples at \url{http://nwis.waterdata.usgs.gov/usa/nwis/qw}.
}
\seealso{
\code{\link{importNWISqw}}
}
\keyword{IO}
\keyword{datasets}

