% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show-censored}
\alias{show,lcens-method}
\alias{show,mcens-method}
\alias{show,qw-method}
\alias{show-censored}
\title{Show Methods for \code{WSqw} objects}
\usage{
\S4method{show}{lcens}(object)

\S4method{show}{mcens}(object)

\S4method{show}{qw}(object)
}
\arguments{
\item{object}{the object to show.}
}
\value{
The object is retruned invisibly.
}
\description{
Prints the object.
}
\keyword{manip}
\keyword{methods}

