\docType{methods}
\name{show-methods}
\alias{show-methods}
\alias{show,lcens-method}
\alias{show,mcens-method}
\alias{show,qw-method}
\title{Show Methods for \code{WSqw} objects}
\description{
  Prints the object.
}
\section{Methods}{
  \describe{

  \item{list("signature(object = \"lcens\")")}{ Prints the
  value and detection limit }

  \item{list("signature(object = \"mcens\")")}{ Prints the
  minimum and maximum values }

  \item{list("signature(object = \"qw\")")}{ Prints the
  minimum and maximum values, and other attributes of the
  data } }
}
\keyword{manip}
\keyword{methods}

