% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/splitQual.R
\name{splitQual}
\alias{splitQual}
\title{Split Leading Qualifiers}
\usage{
splitQual(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{a character vector, missing values are permitted and result in missing values in the output}

\item{name}{a character string giving the names of the output columns.}
}
\value{
A data frame with 2 columns, containing the numeric values and the qualification codes.
}
\description{
Split vector of strings of combined data into numeric and remark columns
}
\examples{
splitQual(c("<1", "1", "<1", "1", "2"), name="Dummy")
}

