\name{vif.censReg}
\alias{vif.censReg}
\title{Variance Inflation Factors}
\usage{
  \method{vif}{censReg} (model, ...)
}
\arguments{
  \item{model}{an object of class "censReg"---output from
  \code{censReg}.}

  \item{\dots}{further arguments passed to or from other
  methods.}
}
\value{
  further arguments passed to or from other methods.
}
\description{
  Computes the variance inflation factor (Helsel and
  Hirsch, 2002) for each variable in an AMLE/MLE
  regression.
}
\seealso{
  \code{\link{censReg}}
}
\keyword{regression}

