% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyPlot-censored.R
\docType{methods}
\name{xyPlot-censored}
\alias{xyPlot,Date,qw-method}
\alias{xyPlot,lcens,lcens-method}
\alias{xyPlot,numeric,lcens-method}
\alias{xyPlot,numeric,mcens-method}
\alias{xyPlot,numeric,qw-method}
\alias{xyPlot-censored}
\title{Censored Data Methods for Function \code{xyPlot}}
\usage{
\S4method{xyPlot}{numeric,lcens}(x, y, Plot = list(name = "", what = "points",
  type = "solid", width = "standard", symbol = "circle", filled = TRUE, size =
  0.09, color = "black"), yaxis.log = FALSE, yaxis.rev = FALSE,
  yaxis.range = c(NA, NA), xaxis.log = FALSE, xaxis.range = c(NA, NA),
  ylabels = 7, xlabels = 7, xtitle = deparse(substitute(x)),
  ytitle = deparse(substitute(y)), caption = "", margin = c(NA, NA, NA,
  NA), Censored = list(name = "", what = "points", symbol = "circle", filled =
  FALSE, size = 0.09, color = "black", width = "hairline", bar = FALSE), ...)

\S4method{xyPlot}{numeric,qw}(x, y, Plot = list(name = "", what = "points",
  type = "solid", width = "standard", symbol = "circle", filled = TRUE, size =
  0.09, color = "black"), yaxis.log = FALSE, yaxis.rev = FALSE,
  yaxis.range = c(NA, NA), xaxis.log = FALSE, xaxis.range = c(NA, NA),
  ylabels = 7, xlabels = 7, xtitle = "",
  ytitle = deparse(substitute(y)), caption = "", margin = c(NA, NA, NA,
  NA), Censored = list(name = "", what = "points", symbol = "circle", filled =
  FALSE, size = 0.09, color = "black", width = "hairline", bar = FALSE), ...)

\S4method{xyPlot}{lcens,lcens}(x, y, Plot = list(name = "", what = "points",
  type = "solid", width = "standard", symbol = "circle", filled = TRUE, size =
  0.09, color = "black"), yaxis.log = FALSE, yaxis.rev = FALSE,
  yaxis.range = c(NA, NA), xaxis.log = FALSE, xaxis.range = c(NA, NA),
  ylabels = 7, xlabels = 7, xtitle = deparse(substitute(x)),
  ytitle = deparse(substitute(y)), caption = "", margin = c(NA, NA, NA,
  NA), Censored = list(name = "", what = "points", symbol = "circle", filled =
  FALSE, size = 0.09, color = "black", width = "hairline", bar = TRUE), ...)

\S4method{xyPlot}{numeric,mcens}(x, y, Plot = list(name = "", what = "points",
  type = "solid", width = "standard", symbol = "circle", filled = TRUE, size =
  0.09, color = "black"), yaxis.log = FALSE, yaxis.rev = FALSE,
  yaxis.range = c(NA, NA), xaxis.log = FALSE, xaxis.range = c(NA, NA),
  ylabels = 7, xlabels = 7, xtitle = deparse(substitute(x)),
  ytitle = deparse(substitute(y)), caption = "", margin = c(NA, NA, NA,
  NA), Censored = list(name = "", what = "points", symbol = "circle", filled =
  FALSE, size = 0.09, color = "black", width = "hairline", bar = TRUE), ...)
}
\arguments{
\item{x}{the x-axis data.}

\item{y}{the y-axis data.}

\item{Plot}{control parameters of the plot for uncensored data.}

\item{yaxis.log}{log-transform the y axis?}

\item{yaxis.rev}{reverse the y axis?}

\item{yaxis.range}{set the range of the y-axis.}

\item{xaxis.log}{logical, if \code{TRUE}, then log-transform the x axis.}

\item{xaxis.range}{set the range of the x-axis.}

\item{ylabels}{set up y-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xlabels}{set up x-axis labels. See \bold{Details}.}

\item{xtitle}{the x-axis title.}

\item{ytitle}{the y-axis title.}

\item{caption}{the figure caption.}

\item{margin}{set up the plot area margins.}

\item{Censored}{control parameters of the plot for censored data. Same arguments as 
for \code{Plot}, but adds \code{bar} which if set to \code{TRUE} will draw lines from
left-censored values to the minimum and draw lines from right-censored values to the
maximum. No bars are drawn for interval censored values.}

\item{...}{additional arguments for specific methods.}
}
\value{
Information about the graph.
}
\description{
Creates a line/scatter plot: selected methods for "lcens," 
"mcens," and "qw" data.

Setting \code{xlabels} to 0 or negative values will 
suppress ticks and labels. If negative, then try to create that absolute value
number of labels. That can be useful for relative axes or specialized labeling,
but is only valid for numeric \code{x}.
}
\note{
A call should be made to \code{setPage} to set up the graphics
environment before calling \code{xyPlot}.
}
\keyword{hplot}
\keyword{methods}

