% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{as.mcens}
\alias{as.mcens}
\alias{as.mcens,lcens,missing,missing-method}
\alias{as.mcens,mcens,missing,missing-method}
\alias{as.mcens,numeric,missing,character-method}
\alias{as.mcens,numeric,missing,missing-method}
\alias{as.mcens,numeric,missing,numeric-method}
\alias{as.mcens,numeric,numeric,missing-method}
\alias{as.mcens,qw,missing,missing-method}
\title{Multiply-censored Data Conversion}
\usage{
as.mcens(lower.val, upper.val, censor.codes)

\S4method{as.mcens}{mcens,missing,missing}(lower.val, upper.val, censor.codes)

\S4method{as.mcens}{lcens,missing,missing}(lower.val, upper.val, censor.codes)

\S4method{as.mcens}{numeric,missing,missing}(lower.val, upper.val, censor.codes)

\S4method{as.mcens}{numeric,numeric,missing}(lower.val, upper.val, censor.codes)

\S4method{as.mcens}{numeric,missing,numeric}(lower.val, upper.val, censor.codes)

\S4method{as.mcens}{numeric,missing,character}(lower.val, upper.val,
  censor.codes)

\S4method{as.mcens}{qw,missing,missing}(lower.val, upper.val, censor.codes)
}
\arguments{
\item{lower.val}{The lower limit of the actual value, the special value of
\code{-Inf} or \code{NA} can be used to indicate left-censoring.}

\item{upper.val}{The upper limit of the actual value, the special value of
\code{Inf} or \code{NA} can be used to indicate right-censoring.}

\item{censor.codes}{optional codes if \code{upper.val} is missing. Any
numeric value less than 0 indicates left-censored, any value greater than 0
indicates right-censored, and 0 indicates an observed value. The character
value "<" indicates left-censored, ">" indicates right-censored, any anything
else indicates an observed value.}
}
\value{
An S4 object of class "mcens." These objects are the basis for the
analysis of censored data that are not strictly left censored.
}
\description{
Convert from various formats to a multiply-censored data (\code{mcens})
object.
}
\details{
In keeping with the logic of \code{Surv}, \code{NA} is permitted to indicate
left- or right-censored data. If both are \code{NA}, then the observation is
treated as missing.
}
\examples{
## Create one of each type of censoring, including uncensored
# the last value is missing
as.mcens(c(-Inf, 2, 2, 5, NA), c(1, 2, 3, Inf, NA))
}
\references{
Lorenz, D.L., in preparation
}
\seealso{
\code{\link[survival]{Surv}}
}
\keyword{manip}
\keyword{methods}

