% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlotCensStats.R
\name{boxPlotCensStats}
\alias{boxPlotCensStats}
\title{Not exported}
\usage{
boxPlotCensStats(x, Box, yaxis.log)
}
\arguments{
\item{x}{list of the lcens data}

\item{Box}{the Box parameters}

\item{yaxis.log}{is it?}
}
\value{
Info to actually plot the boxplot.
}
\description{
Computes the statistics for a left-censored data boxplot.
}

