% Generated by roxygen2 (4.0.2): do not edit by hand
\name{c.lcens}
\alias{c.lcens}
\alias{c.mcens}
\alias{c.qw}
\title{Combine Values into a Vector}
\usage{
\method{c}{lcens}(..., recursive = FALSE)

\method{c}{mcens}(..., recursive = FALSE)

\method{c}{qw}(..., recursive = FALSE)
}
\arguments{
\item{recursive}{not used.}

\item{\dots}{any objects that can be coerced to the class of the first one.}
}
\value{
A vector of an appropriate class.
}
\description{
Combines arguments to form a vector for censored or water-quality data:
methods for "cens," "mcens," and "qw" data. All arguments are coerced to
a common type which is the type of the first argument, except for the qw method,
}
\examples{
c(as.lcens(c(1,3), 2), 2)
}

