% Generated by roxygen2 (4.0.2): do not edit by hand
\name{censReg}
\alias{censReg}
\title{Censored Data Regression}
\usage{
censReg(formula, data, subset, weights, na.action, dist = "normal")
}
\arguments{
\item{formula}{a formula describing the regression model. See \bold{Details}.}

\item{data}{the data to search for the variables in \code{formula}.}

\item{subset}{an expression to select a subset of the data.}

\item{weights}{a variable to use for weights.}

\item{na.action}{what to do with missing values.}

\item{dist}{the distribution of the data, either "normal," "lognormal,"
or "commonlog." See \bold{Details}.}
}
\value{
An object of class "censReg."
}
\description{
Build a regression model for censored response data.
}
\details{
The left-hand side of the formula may be any numeric variable, just as with
\code{lm} or a variable of class "lcens," "mcens," or "qw."\cr
For un- or left-censored data, AMLE is used unless weights are specified in
the model, then MLE is used, through a call to survreg. For any other
censored data, MLE is used.\cr
If \code{dist} is "normal," then the regression analysis assumes that the
residuals are normally distributed. If \code{dist} is "lognormal," then
the regression analysis assumes that the residuals are lognormally
distributed. In this case, predicted values are back transformed and
optionally bias corrected to represent the expected mean. If \code{dist}
is "commonlog," then the response data are transformed using \code{log10}
and those residuals are assumed to be normally distributed. No back
transformation is made for predicted values.
}
\note{
A special feature of \code{censReg} is the ability to expand
the formula passed as a variable rather than as an expressly entered
formula. This feature is intended to facilitate certain kinds of scripts that
construct a formula from combinations of vairabales in a dataset.
}
\examples{
set.seed(345)
X <- runif(24, 1, 5)
Y <- X/2 + rnorm(24)
lm(Y ~ X) # the uncensored regression
censReg(as.lcens(Y, 1) ~ X) # censored at 1
}
\references{
Lorenz, 2015, smwrQW.\cr
Breen, R., 1996, Regression models: censored, sample selected, or truncated data:
Sage University Paper series on Qunatitative Applications in the Social Sciences,
07-111, Thousand Oaks, CA,
Cohn, T.A., 1988, Adjusted maximum likelihood estimation of moments
of lognormal populations from type I censored samples:
U.S. Geological Survey Open-File Report 88-350, 34 p.\cr
#'
}
\seealso{
\code{\link{lm}}, \code{\link{survreg}}
}
\keyword{censored}
\keyword{regression}

