% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fillIn}
\alias{fillIn}
\title{Estimate Censored Values}
\usage{
fillIn(x, method = "ROS", alpha = 0.4)
}
\arguments{
\item{x}{an object of class "lcens." Missing values are ignored.}

\item{method}{the method to use: "ROS," "log ROS," "MLE," "log MLE," or
"triangular." The method of Regression on Order Statistics and MLE is
explained in Helsel (2012). The "log ROS" first log-transform the data and
back-transforms the estimated values. The traingular method distributes the
censored value assuming a traingular distrib ution between 0 and the single
detectgion limit.}

\item{alpha}{the offset fraction to be used; typically in [0,0.5].}
}
\value{
A vector of sorted estiamtes and actual values.
}
\description{
Estimates values of left-censored data
}
\examples{
set.seed(5420)
XR <- sort(rnorm(10))
XF <- fillIn(as.lcens(XR, -1)) # censors lowest 3 values
#How'd we do?
rbind(XR, XF)
#Note that this is unusual becuase all 10 random values were less than the mean!
}
\references{
Helsel, D.R. 2012, Blah blah: some place, Wiley, some pages.\cr
}
\seealso{
\code{\link{as.double.qw}}
}
\keyword{manip}

