% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.censReg.R
\name{rmse.censReg}
\alias{rmse.censReg}
\title{Root-Mean-Squared Error}
\usage{
\method{rmse}{censReg}(x, ...)
}
\arguments{
\item{x}{the output from \code{censReg}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
The estimated root-mean-squared error, also know as the residual
standard error.
}
\description{
Computes the root-mean-squared error (RMSE) of a censored regression object: 
method for "censReg" object.
}
\seealso{
\code{\link{censReg}}
}

