% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vif.censReg.R
\name{vif.censReg}
\alias{vif.censReg}
\title{Variance Inflation Factors}
\usage{
\method{vif}{censReg}(model, ...)
}
\arguments{
\item{model}{an object of class "censReg"---output from \code{censReg}.}

\item{\dots}{not used, further arguments passed to or from other methods.}
}
\value{
A named numeric vector containing the variance inflation factors
for each variable.
}
\description{
Computes the variance inflation factor (Helsel and Hirsch, 2002) for each
variable in an AMLE/MLE regression.
}
\references{
Helsel, D.R. and Hirsch, R.M., 2002, Statistical methods in
water resources: U.S. Geological Survey Techniques of Water-Resources
Investigations, book 4, chap. A3, 522 p.
}
\seealso{
\code{\link{censReg}}
}
\keyword{regression}

